requirements.txt
streamlit
python-dotenv
httpx
langchain
langchain_openai